requirements.txt
streamlit>=1.38.0
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
duckdb>=1.0.0
pandas>=2.0.0
PyMuPDF>=1.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
Pillow>=10.0.0
openpyxl>=3.1.0
lxml>=5.0.0
python-dotenv>=1.0.0
python-dateutil>=2.8.0
imagehash>=4.3.0
pytesseract>=0.3.10
opencv-python>=4.8.0
